% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calcul de la distribution des âges des élus municipaux}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un DataFrame contenant les informations des élus municipaux. Il doit inclure une colonne Date.de.naissance au format jj/mm/aaaa.}
}
\value{
Un vecteur numérique contenant les quantiles de l'âge : minimum, premier quartile (Q1), médiane (Q2), troisième quartile (Q3) et maximum.
}
\description{
Cette fonction calcule les quantiles de la distribution des âges des élus municipaux dans un DataFrame.
}
\details{
\itemize{
\item La fonction convertit la colonne \code{Date.de.naissance} en format Date.
\item Elle calcule l'âge en années à partir de la date de naissance.
\item Elle retourne les quantiles 0\% (min), 25\% (Q1), 50\% (médiane), 75\% (Q3) et 100\% (max).
\item Les valeurs manquantes (\code{NA}) sont ignorées dans le calcul.
}
}
