% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Création d'un objet spécifique pour les départements}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un dataframe contenant les informations suivantes :
\itemize{
\item \code{Code.du.département} : Code du département.
\item \code{Libellé.du.département} : Nom du département.
\item \code{Code.de.la.commune} : Code de la commune.
\item \code{Libellé.de.la.commune} : Nom de la commune.
\item \code{Code.de.la.collectivité.à.statut.particulier} (optionnel) : Code de la collectivité à statut particulier.
\item \code{Libellé.de.la.collectivité.à.statut.particulier} (optionnel) : Nom de la collectivité à statut particulier.
}}
}
\value{
Un objet de type \code{departement}, qui est une liste contenant les informations suivantes :
\itemize{
\item \code{Code.du.département} : Code unique du département.
\item \code{Libellé.du.département} : Nom du département.
\item \code{Collectivite} : Liste des collectivités à statut particulier si elles existent dans les données (sinon \code{NULL}).
\item \code{Communes} : Liste des objets de type \code{commune} créés pour chaque commune présente dans le département.
}
}
\description{
Cette fonction permet de créer un objet de type departement à partir d'un dataframe contenant des informations sur les communes, les élus et les collectivités. Elle extrait des informations essentielles concernant un département y compris les communes qui le composent, les élus et leurs âges, ainsi que des informations supplémentaires sur les collectivités à statut particulier.
}
