% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport_pdf.R
\name{generer_rapport_pdf}
\alias{generer_rapport_pdf}
\title{Générer un rapport pour une commune et un département en PDF}
\usage{
generer_rapport_pdf(commune, departement, output)
}
\arguments{
\item{commune}{Un code de commune sous forme de chaîne de caractères (ex : "44109").}

\item{departement}{Un code de département sous forme de chaîne de caractères (ex : "44").}

\item{output}{Chemin de sortie du fichier PDF généré, sous forme de chaîne de caractères.}
}
\value{
un fichier PDF
}
\description{
Cette fonction génère un rapport PDF à partir d'un fichier Quarto. Elle utilise les codes de la commune et du département spécifiés, puis enregistre le rapport dans le dossier de sortie.
}
\examples{
\dontrun{
generer_rapport_pdf(
 commune = "44330",
 departement = "44",
 output = "rapport_la_chapelle.pdf"
)
}

}
