% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.commune.R
\name{plot.commune}
\alias{plot.commune}
\title{Visualisation de la répartition des élus par catégorie socio-professionnelle pour une commune}
\usage{
\method{plot}{commune}(df)
}
\arguments{
\item{df}{Un dataframe contenant les informations des élus municipaux d'une seule commune.
Il doit inclure au minimum les colonnes suivantes :
\itemize{
\item \code{Libellé.de.la.commune} : Nom de la commune.
\item \code{Libellé.du.département} : Nom du département.
\item \code{Code.de.la.catégorie.socio.professionnelle} : Code représentant la catégorie socio-professionnelle des élus.
\item \code{Nom.de.l.élu} : Nom des élus.
}}
}
\value{
Un graphique en barres affichant le nombre d'élus par code professionnel pour la commune sélectionnée.
}
\description{
Cette fonction génère un graphique en barres représentant la répartition des élus municipaux en fonction de leur catégorie socio-professionnelle pour une commune donnée.
}
