% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Cette fonction compte pour chaque code professionnel, le nombre d’élus ayant ce code}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un dataframe contenant les données des élus, avec une colonne nommée \code{Code.de.la.catégorie.socio.professionnelle} qui représente les codes professionnels des élus.}
}
\value{
Cette fonction génère un graphique à barres horizontal affichant le nombre d'élus par code professionnel. Le graphique est tracé directement dans la sortie graphique de R sans nécessiter de librairies externes.
}
\description{
Cette fonction calcule, pour chaque code professionnel dans un jeu de données, le nombre d'élus ayant ce code, puis génère un graphique à barres horizontal pour afficher la répartition des élus par code professionnel.
}
