% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.departement.R
\name{summary.departement}
\alias{summary.departement}
\title{Résumé des caractéristiques d'un département}
\usage{
\method{summary}{departement}(x)
}
\arguments{
\item{x}{Un objet de classe département, contenant les informations des élus municipaux du département.}
}
\value{
Un résumé textuel affiché dans la console avec les statistiques sur les élus du département.
}
\description{
Cette fonction fournit un résumé des informations concernant les élus municipaux d'un département.
Elle affiche plusieurs statistiques clés, notamment :
\itemize{
\item Le nom du département.
\item Le nombre total de communes et d'élus.
\item La distribution des âges des élus.
\item L'élu.e le/la plus âgé.e et le/la plus jeune, ainsi que leur commune.
\item La commune avec la moyenne d'âge la plus faible et celle avec la moyenne la plus élevée.
}
}
\details{
\itemize{
\item La fonction vérifie d'abord que x est bien un objet de classe "département".
\item Elle calcule les statistiques en utilisant les dates de naissance pour déterminer l'âge des élus.
\item Les résultats sont affichés directement dans la console avec print().
}
}
