% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.commune.R
\name{summary_commune}
\alias{summary_commune}
\alias{summary.commune}
\title{Résumé des caractéristiques des élus municipaux d'une commune}
\usage{
\method{summary}{commune}(x)
}
\arguments{
\item{x}{Un objet de classe \code{commune} contenant les informations des élus municipaux.
Il doit inclure les colonnes \code{Libellé.de.la.commune}, \code{Nom.de.l.élu}, \code{Prénom.de.l.élu} et \code{Âge}.}
}
\value{
Un affichage des informations clés sur les élus municipaux de la commune
}
\description{
Cette fonction fournit un résumé des caractéristiques des élus municipaux d'une commune donnée.
Elle affiche le nom de la commune, le nombre total d'élus, la distribution des âges des élus,
ainsi que les informations de l'élu(e) le plus âgé(e).
}
