% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu municipal le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un DataFrame contenant les informations des élus municipaux. Il doit inclure une colonne Date.de.naissance au format jj/mm/aaaa.}
}
\value{
Un DataFrame avec le nom, le prénom, la date de naissance et l'âge de l'élu le plus âgé.
}
\description{
Cette fonction identifie l'élu le plus âgé dans un DataFrame contenant des élus municipaux.
}
\details{
La fonction convertit la colonne Date.de.naissance en format Date, identifie la plus ancienne date de naissance et calcule l'âge en années.
}
\seealso{
\code{\link[=validate_schema]{validate_schema()}}
}
